refactor(engine): extract WS health watchdog from TradingEngine
strategy.py carried the ~218-line _monitor_ws_health watchdog (WS staleness
monitoring, reconnect/backoff, degradation escalation) as an engine method. Move
it into engines/ws_health_monitor.py as monitor_ws_health(engine) (2594->2378),
keeping a thin delegator so watchdog task creation and tests are unchanged.

Typed engine as TradingEngine. Behavior-preserving: 125 ws-health/chaos tests +
the full 1225 live_trade suite pass. Completes the three large engine clusters
(reconciliation, order placement, WS health); strategy.py 3538->2378 (-33%).
